refactor: centralize issue ts alias policy
